# internal/docs/swagger.yaml
basePath: /
definitions:
  api.authResult:
    properties:
      authenticated:
        type: boolean
    type: object
...
  api.bandwidthUpdate:
    properties:
      bandwidth:
        description: e.g. "2M"
        type: string
      halfSaturation:
        description: leechers for half speed
        type: number
    type: object
...
  api.errorResponse:
    properties:
      error:
        type: string
    type: object
  api.inviteBody:
    properties:
      invite:
        type: string
    type: object
  api.inviteRequest:
    properties:
      username:
        type: string
    type: object
  api.inviteResponse:
    properties:
      code:
        type: string
      url:
        type: string
    type: object
...
  api.proxyInfo:
    properties:
      default:
        description: server default (--client.proxy), "" if none
        type: string
      error:
        description: probe error when status is "error"
        type: string
      proxy:
        description: effective proxy URL ("" = announce directly)
        type: string
      status:
        description: '"ok" | "error" | "direct" | "unknown"'
        type: string
    type: object
  api.proxyUpdate:
    properties:
      proxy:
        type: string
    type: object
  api.registerChallenge:
    properties:
      ceremonyId:
        type: string
      options:
        $ref: '#/definitions/protocol.CredentialCreation'
      username:
        type: string
    type: object
...
  protocol.AttestationFormat:
    enum:
    - packed
    - tpm
    - android-key
    - android-safetynet
    - fido-u2f
    - apple
    - compound
    - none
    type: string
    x-enum-varnames:
    - AttestationFormatPacked
    - AttestationFormatTPM
    - AttestationFormatAndroidKey
    - AttestationFormatAndroidSafetyNet
    - AttestationFormatFIDOUniversalSecondFactor
    - AttestationFormatApple
    - AttestationFormatCompound
    - AttestationFormatNone
  protocol.AuthenticationExtensions:
    additionalProperties: {}
    type: object
  protocol.AuthenticatorAttachment:
    enum:
    - platform
    - cross-platform
    type: string
    x-enum-varnames:
    - Platform
    - CrossPlatform
  protocol.AuthenticatorSelection:
    properties:
      authenticatorAttachment:
        allOf:
        - $ref: '#/definitions/protocol.AuthenticatorAttachment'
        description: |-
          AuthenticatorAttachment If this member is present, eligible authenticators are filtered to only
          authenticators attached with the specified AuthenticatorAttachment enum.
      requireResidentKey:
        description: |-
          RequireResidentKey this member describes the Relying Party's requirements regarding resident
          credentials. If the parameter is set to true, the authenticator MUST create a client-side-resident
          public key credential source when creating a public key credential.
        type: boolean
      residentKey:
        allOf:
        - $ref: '#/definitions/protocol.ResidentKeyRequirement'
        description: |-
          ResidentKey this member describes the Relying Party's requirements regarding resident
          credentials per Webauthn Level 2.
      userVerification:
        allOf:
        - $ref: '#/definitions/protocol.UserVerificationRequirement'
        description: |-
          UserVerification This member describes the Relying Party's requirements regarding user verification for
          the create() operation. Eligible authenticators are filtered to only those capable of satisfying this
          requirement.
    type: object
...
  protocol.ConveyancePreference:
    enum:
    - none
    - indirect
    - direct
    - enterprise
    type: string
    x-enum-varnames:
    - PreferNoAttestation
    - PreferIndirectAttestation
    - PreferDirectAttestation
    - PreferEnterpriseAttestation
...
  protocol.CredentialCreation:
    properties:
      mediation:
        $ref: '#/definitions/protocol.CredentialMediationRequirement'
      publicKey:
        $ref: '#/definitions/protocol.PublicKeyCredentialCreationOptions'
    type: object
  protocol.CredentialDescriptor:
    properties:
      id:
        description: CredentialID The ID of a credential to allow/disallow.
        items:
          type: integer
        type: array
      transports:
        description: The authenticator transports that can be used.
        items:
          $ref: '#/definitions/protocol.AuthenticatorTransport'
        type: array
      type:
        allOf:
        - $ref: '#/definitions/protocol.CredentialType'
        description: The valid credential types.
    type: object
  protocol.CredentialMediationRequirement:
    enum:
    - ""
    - silent
    - optional
    - conditional
    - required
    type: string
    x-enum-varnames:
    - MediationDefault
    - MediationSilent
    - MediationOptional
    - MediationConditional
    - MediationRequired
  protocol.CredentialParameter:
    properties:
      alg:
        $ref: '#/definitions/webauthncose.COSEAlgorithmIdentifier'
      type:
        $ref: '#/definitions/protocol.CredentialType'
    type: object
...
  protocol.PublicKeyCredentialCreationOptions:
    properties:
      attestation:
        $ref: '#/definitions/protocol.ConveyancePreference'
      attestationFormats:
        items:
          $ref: '#/definitions/protocol.AttestationFormat'
        type: array
      authenticatorSelection:
        $ref: '#/definitions/protocol.AuthenticatorSelection'
      challenge:
        items:
          type: integer
        type: array
      excludeCredentials:
        items:
          $ref: '#/definitions/protocol.CredentialDescriptor'
        type: array
      extensions:
        $ref: '#/definitions/protocol.AuthenticationExtensions'
      hints:
        items:
          $ref: '#/definitions/protocol.PublicKeyCredentialHints'
        type: array
      pubKeyCredParams:
        items:
          $ref: '#/definitions/protocol.CredentialParameter'
        type: array
      rp:
        $ref: '#/definitions/protocol.RelyingPartyEntity'
      timeout:
        type: integer
      user:
        $ref: '#/definitions/protocol.UserEntity'
    type: object
  protocol.PublicKeyCredentialHints:
    enum:
    - security-key
    - client-device
    - hybrid
    type: string
    x-enum-varnames:
    - PublicKeyCredentialHintSecurityKey
    - PublicKeyCredentialHintClientDevice
    - PublicKeyCredentialHintHybrid
...
  protocol.RelyingPartyEntity:
    properties:
      id:
        description: A unique identifier for the Relying Party entity, which sets
          the RP ID.
        type: string
      name:
        description: |-
          A human-palatable name for the entity. Its function depends on what the PublicKeyCredentialEntity represents:

          When inherited by PublicKeyCredentialRpEntity it is a human-palatable identifier for the Relying Party,
          intended only for display. For example, "ACME Corporation", "Wonderful Widgets, Inc." or "ОАО Примертех".

          When inherited by PublicKeyCredentialUserEntity, it is a human-palatable identifier for a user account. It is
          intended only for display, i.e., aiding the user in determining the difference between user accounts with similar
          displayNames. For example, "alexm", "alex.p.mueller@example.com" or "+14255551234".
        type: string
    type: object
  protocol.ResidentKeyRequirement:
    enum:
    - discouraged
    - preferred
    - required
    type: string
    x-enum-varnames:
    - ResidentKeyRequirementDiscouraged
    - ResidentKeyRequirementPreferred
    - ResidentKeyRequirementRequired
  protocol.UserEntity:
    properties:
      displayName:
        description: |-
          A human-palatable name for the user account, intended only for display.
          For example, "Alex P. Müller" or "田中 倫". The Relying Party SHOULD let
          the user choose this, and SHOULD NOT restrict the choice more than necessary.
        type: string
      id:
        description: |-
          ID is the user handle of the user account entity. To ensure secure operation,
          authentication and authorization decisions MUST be made on the basis of this id
          member, not the displayName nor name members. See Section 6.1 of
          [RFC8266](https://www.w3.org/TR/webauthn/#biblio-rfc8266).
      name:
        description: |-
          A human-palatable name for the entity. Its function depends on what the PublicKeyCredentialEntity represents:

          When inherited by PublicKeyCredentialRpEntity it is a human-palatable identifier for the Relying Party,
          intended only for display. For example, "ACME Corporation", "Wonderful Widgets, Inc." or "ОАО Примертех".

          When inherited by PublicKeyCredentialUserEntity, it is a human-palatable identifier for a user account. It is
          intended only for display, i.e., aiding the user in determining the difference between user accounts with similar
          displayNames. For example, "alexm", "alex.p.mueller@example.com" or "+14255551234".
        type: string
    type: object
...
info:
  contact: {}
  description: Spoof BitTorrent tracker announces to boost your ratio.
  title: emission API
  version: "1.0"
paths:
  /api/auth/credentials/{id}:
    delete:
      parameters:
      - description: Base64url credential ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Remove a passkey (admin only)
      tags:
      - admin
  /api/auth/invite:
    post:
      consumes:
      - application/json
      parameters:
      - description: Username for the invite
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.inviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.inviteResponse'
      summary: Create an invite link
      tags:
      - auth
...
  /api/auth/invites/{token}:
    delete:
      parameters:
      - description: Invite token
        in: path
        name: token
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Revoke a pending invite (admin only)
      tags:
      - admin
...
  /api/auth/login/finish:
    post:
      consumes:
      - application/json
      parameters:
      - description: Ceremony ID from /begin
        in: query
        name: ceremony
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.authResult'
      summary: Finish passkey login
      tags:
      - auth
...
  /api/auth/me/devices/{id}:
    delete:
      parameters:
      - description: Base64url credential ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Remove one of my devices
      tags:
      - auth
  /api/auth/register/begin:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invite token (empty string for bootstrap)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.inviteBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.registerChallenge'
      summary: Begin passkey registration
      tags:
      - auth
  /api/auth/register/finish:
    post:
      consumes:
      - application/json
      parameters:
      - description: Ceremony ID from /begin
        in: query
        name: ceremony
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.authResult'
      summary: Finish passkey registration
      tags:
      - auth
...
  /api/auth/users/{username}:
    delete:
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Remove a user and their torrents (admin only)
      tags:
      - admin
  /api/auth/users/{username}/bandwidth:
    put:
      consumes:
      - application/json
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Bandwidth (e.g. 2M)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.bandwidthUpdate'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Set a user's upload bandwidth (admin only)
      tags:
      - admin
  /api/auth/users/{username}/proxy:
    get:
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.proxyInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Get a user's tracker proxy (admin only)
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Proxy URL (empty = direct)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.proxyUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.proxyInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Set a user's tracker proxy (admin only)
      tags:
      - admin
...
swagger: "2.0"